requirements.txt
pylsl
requests
paho-mqtt
# RPi.GPIO       # Only used on Raspberry Pi (safe to install elsewhere)
pytest